wip: callfn
